You can use immutable data as well
But you need to juggle a bit with it.
I think the .d.ts files are a bit wrong, I started to play with them to get them working.
I exported defaults, probably those were not part of the typescript version they were using.

You can't import Map in the way it's declared, I'll have to learn how to declare it properly.